Section 3, Vid 113 - Loading & Rendering Meal Details via Dynamic Routes & Route Parameters
Populated content onto our MealDetailPage component with help form the getMeal function inside of lib/meals.js . This function has a select statement that returns all meals where the slug of the meal = the slug that is extracted from the end portion from the url to return a specific meal. This is done because of the dynamic routing we have setup previously.