Section 3, Vid 114 - Throwing Not Found Errors for Individual Meals & Vid 115 - Getting Started with the 'Share Meal' Form
Used the special notFound function from react that will automatically render the closest not-found.js or error.js file. In our case we want to render the closest not-found.js file when a meal was not returned from the database. Added a page.js and page.module.css file provided from the lecture to help setup our Share Form.