Section 3, Vid 116 - Getting Started with a Custom Image Picker Input Component
Added a image-picker.js file that will contain a ImagePicker component that will render an input field that should allow us to upload images. We hid this input to allow us to have a more stylish button that will have a click event that will eventually reference the input field.