Section 10, Vid 241 - Running MongoDB Queries From Inside API Routes
Setup MongoDB in our environment by download the necessary pages and connecting to the database we have setup on the mongoDB website. We created a 'newsletter' database in the cluster, and created a 'emails' collection where we will store our emails of the users that wish to register to the newsletter.